+++ pp/test/test_netlists.py
@@ -21,15 +21,32 @@ def test_netlists_instances(component_type, data_regression):
     data_regression.check(d)
 
 
+def demo_netlist(component_type):
+    c1 = component_factory[component_type]()
+    n = c1.get_netlist()
+    yaml_str = OmegaConf.to_yaml(n, sort_keys=True)
+    c2 = pp.component_from_yaml(yaml_str)
+    pp.show(c2)
+
+
 if __name__ == "__main__":
     # c = component_factory["mzi"]()
-    # c = component_factory["ring_single"]()
     # c = component_factory["ring_double"]()
 
-    pp.clear_connections()
-    c = component_factory["ring_double"]()
-    n = c.get_netlist()
-    print(n.connections)
+    # pp.clear_connections()
+    # c = component_factory["ring_double"]()
+    # n = c.get_netlist()
+    # print(n.connections)
     # n = c.get_netlist_yaml()
     # print(n)
-    pp.show(c)
+    # pp.show(c)
+
+    # c = component_factory["ring_single"]()
+    component_type = "ring_single"
+    c1 = component_factory[component_type]()
+    n = c1.get_netlist()
+    pp.clear_cache()
+    yaml_str = OmegaConf.to_yaml(n, sort_keys=True)
+    print(yaml_str)
+    c2 = pp.component_from_yaml(yaml_str)
+    pp.show(c2)